requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cfgv==3.5.0
charset-normalizer==3.4.7
click==8.3.3
distlib==0.4.0
Django==6.0.4
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.17.1
filelock==3.29.0
gunicorn==25.3.0
identify==2.6.19
idna==3.11
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.1
pathspec==1.1.1
platformdirs==4.9.6
pre_commit==4.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.12.1
python-discovery==1.2.2
python-dotenv==1.2.2
pytokens==0.4.1
PyYAML==6.0.3
requests==2.33.1
ruff==0.15.12
sqlparse==0.5.5
urllib3==2.6.3
uuid6==2025.0.1
virtualenv==21.3.0
